Revive bsatool
