+++ components/bsa/tests/Makefile
@@ -1,6 +1,6 @@
 GCC=g++
 
-all: bsa_file_test bsatool ogre_archive_test
+all: bsa_file_test ogre_archive_test
 
 I_OGRE=$(shell pkg-config --cflags OGRE)
 L_OGRE=$(shell pkg-config --libs OGRE)
@@ -11,12 +11,5 @@ bsa_file_test: bsa_file_test.cpp ../bsa_file.cpp
 ogre_archive_test: ogre_archive_test.cpp ../bsa_file.cpp ../bsa_archive.cpp
 	$(GCC) $^ -o $@ $(I_OGRE) $(L_OGRE)
 
-bsatool: bsatool.cpp ../bsa_file.cpp bsatool_cmd.c
-	$(GCC) $^ -o $@
-
-bsatool_cmd.c: bsatool.ggo
-	gengetopt < bsatool.ggo
-
 clean:
 	rm *_test
-	rm bsatool
